requirements.txt
pyOpenSSL==23.2.0
matplotlib==3.7.1
numpy==1.24.3